+++ examples/unlock_signal.rs
@@ -1,4 +1,8 @@
-use std::{sync::{Arc, Mutex}, thread::sleep, time::Duration};
+use std::{
+    sync::{Arc, Mutex},
+    thread::sleep,
+    time::Duration,
+};
 
 use logind_zbus::{ManagerProxy, SessionProxy};
 use zbus::{Connection, SignalReceiver};
@@ -8,17 +12,17 @@ fn print_unlocked() -> std::result::Result<(), zbus::Error> {
     Ok(())
 }
 
-fn main() {
-    let connection = Connection::new_system().unwrap();
-    let manager = ManagerProxy::new(&connection).unwrap();
-    let sessions = manager.list_sessions().unwrap();
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let connection = Connection::new_system()?;
+    let manager = ManagerProxy::new(&connection)?;
+    let sessions = manager.list_sessions()?;
     dbg!(&sessions);
-    let session = SessionProxy::new(&connection, &sessions[0]).unwrap();
+    let session = SessionProxy::new(&connection, &sessions[0])?;
 
     let end = Arc::new(Mutex::new(false));
-    let end2= end.clone();
+    let end2 = end.clone();
 
-    session.get_proxy().connect_unlock(print_unlocked).unwrap();
+    session.get_proxy().connect_unlock(print_unlocked)?;
 
     session.connect_unlock(move || {
         println!("Session unlocked");
@@ -26,13 +30,13 @@ fn main() {
             *lock = true;
         }
         Ok(())
-    }).unwrap();
+    })?;
 
     let mut signals = SignalReceiver::new(connection);
-    signals.receive_for(session.get_proxy()).unwrap();
+    signals.receive_for(session.get_proxy())?;
 
     loop {
-        signals.next_signal().unwrap();
+        signals.next_signal()?;
         if let Ok(lock) = end.lock() {
             if *lock {
                 break;
@@ -40,4 +44,6 @@ fn main() {
         }
         sleep(Duration::from_millis(100));
     }
+
+    Ok(())
 }