+++ src/proxy/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod manager;
-pub(crate) mod session;
 pub(crate) mod seat;
+pub(crate) mod session;
 pub(crate) mod user;