+++ src/types.rs
@@ -7,6 +7,7 @@ use zvariant_derive::Type;
 
 /// If `IsSupported::Invalid` then the response from
 /// logind was not well defined.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
 pub enum IsSupported {
     NA,
     Yes,
@@ -21,13 +22,14 @@ impl From<&str> for IsSupported {
             "na" => Self::NA,
             "yes" => Self::Yes,
             "no" => Self::No,
-            _ => Self::NA
+            _ => Self::NA,
         }
     }
 }
 
 /// If `ShutdownType::Invalid` then the response from
 /// logind was not well defined.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
 pub enum ShutdownType {
     PowerOff,
     DryPowerOff,
@@ -68,6 +70,7 @@ impl From<ShutdownType> for &str {
 
 /// State of a session. If `SessionState::Invalid` then the response from
 /// logind was not well defined.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
 pub enum SessionState {
     Online,
     Active,
@@ -81,37 +84,14 @@ impl From<&str> for SessionState {
             "online" => Self::Online,
             "active" => Self::Active,
             "closing" => Self::Closing,
-            _ => Self::Invalid
-        }
-    }
-}
-
-/// State of a User. If `UserState::Invalid` then the response from
-/// logind was not well defined.
-pub enum UserState {
-    Online,
-    Offline,
-    Lingering,
-    Active,
-    Closing,
-    Invalid,
-}
-
-impl From<&str> for UserState {
-    fn from(s: &str) -> Self {
-        match s.trim() {
-            "online" => Self::Online,
-            "offline" => Self::Offline,
-            "lingering" => Self::Lingering,
-            "active" => Self::Active,
-            "closing" => Self::Closing,
-            _ => Self::Invalid
+            _ => Self::Invalid,
         }
     }
 }
 
 /// Class of Session. If `SessionClass::Invalid` then the response from
 /// logind was not well defined.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
 pub enum SessionClass {
     User,
     Greeter,
@@ -125,7 +105,7 @@ impl From<&str> for SessionClass {
             "user" => Self::User,
             "greeter" => Self::Greeter,
             "lock-screen" => Self::LockScreen,
-            _ => Self::Invalid
+            _ => Self::Invalid,
         }
     }
 }
@@ -144,6 +124,11 @@ impl Device {
     pub fn inactive(&self) -> bool {
         self.inactive
     }
+}
+
+pub trait IntoSessionPath {
+    fn into_path(&self) -> OwnedObjectPath;
+    fn into_path_ref(&self) -> &OwnedObjectPath;
 }
 
 #[derive(Debug, Type, Serialize, Deserialize)]
@@ -177,6 +162,80 @@ impl SessionInfo {
 
     pub fn path(&self) -> &OwnedObjectPath {
         &self.path
+    }
+}
+
+impl IntoSessionPath for SessionInfo {
+    fn into_path(&self) -> OwnedObjectPath {
+        self.path.clone()
+    }
+
+    fn into_path_ref(&self) -> &OwnedObjectPath {
+        &self.path
+    }
+}
+
+#[derive(Debug, PartialEq, Clone, Type, Serialize, Deserialize)]
+pub struct SessionPath {
+    id: String,
+    /// Name of session user
+    path: OwnedObjectPath,
+}
+
+impl SessionPath {
+    pub fn id(&self) -> &str {
+        &self.id
+    }
+
+    pub fn path(&self) -> &OwnedObjectPath {
+        &self.path
+    }
+}
+
+impl TryFrom<OwnedValue> for SessionPath {
+    type Error = zbus::Error;
+
+    fn try_from(value: OwnedValue) -> Result<Self, Self::Error> {
+        let value = <Structure>::try_from(value)?;
+        return Ok(Self {
+            id: <String>::try_from(value.fields()[0].clone())?,
+            path: <OwnedObjectPath>::try_from(value.fields()[1].clone())?,
+        });
+    }
+}
+
+impl IntoSessionPath for SessionPath {
+    fn into_path(&self) -> OwnedObjectPath {
+        self.path.clone()
+    }
+
+    fn into_path_ref(&self) -> &OwnedObjectPath {
+        &self.path
+    }
+}
+
+/// The type of Session. If `State::Invalid` then the response from
+/// logind was not well defined.
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
+pub enum SessionType {
+    X11,
+    Wayland,
+    MIR,
+    TTY,
+    Unspecified,
+    Invalid,
+}
+
+impl From<&str> for SessionType {
+    fn from(s: &str) -> Self {
+        match s {
+            "wayland" => SessionType::Wayland,
+            "x11" => SessionType::X11,
+            "mir" => SessionType::MIR,
+            "tty" => SessionType::TTY,
+            "unspecified" => SessionType::Unspecified,
+            _ => SessionType::Invalid,
+        }
     }
 }
 
@@ -238,7 +297,41 @@ impl TryFrom<OwnedValue> for DbusPath {
     }
 }
 
-#[derive(Debug, Type, Serialize, Deserialize)]
+#[derive(Debug, PartialEq, Clone, Type, Serialize, Deserialize)]
+pub struct SeatPath {
+    id: String,
+    /// Name of session user
+    path: OwnedObjectPath,
+}
+
+impl SeatPath {
+    pub fn id(&self) -> &str {
+        &self.id
+    }
+
+    pub fn path(&self) -> &OwnedObjectPath {
+        &self.path
+    }
+}
+
+impl TryFrom<OwnedValue> for SeatPath {
+    type Error = zbus::Error;
+
+    fn try_from(value: OwnedValue) -> Result<Self, Self::Error> {
+        let value = <Structure>::try_from(value)?;
+        return Ok(Self {
+            id: <String>::try_from(value.fields()[0].clone())?,
+            path: <OwnedObjectPath>::try_from(value.fields()[1].clone())?,
+        });
+    }
+}
+
+pub trait IntoUserPath {
+    fn into_path(&self) -> OwnedObjectPath;
+    fn into_path_ref(&self) -> &OwnedObjectPath;
+}
+
+#[derive(Debug, Clone, Type, Serialize, Deserialize)]
 pub struct UserInfo {
     uid: u32,
     name: String,
@@ -260,14 +353,24 @@ impl UserInfo {
     }
 }
 
-#[derive(Debug, Type, Serialize, Deserialize)]
-pub struct UserSelf {
+impl IntoUserPath for UserInfo {
+    fn into_path(&self) -> OwnedObjectPath {
+        self.path.clone()
+    }
+
+    fn into_path_ref(&self) -> &OwnedObjectPath {
+        &self.path
+    }
+}
+
+#[derive(Debug, Clone, Type, Serialize, Deserialize)]
+pub struct UserPath {
     uid: u32,
     /// Name of session user
     path: OwnedObjectPath,
 }
 
-impl UserSelf {
+impl UserPath {
     pub fn uid(&self) -> u32 {
         self.uid
     }
@@ -277,7 +380,7 @@ impl UserSelf {
     }
 }
 
-impl TryFrom<OwnedValue> for UserSelf {
+impl TryFrom<OwnedValue> for UserPath {
     type Error = zbus::Error;
 
     fn try_from(value: OwnedValue) -> Result<Self, Self::Error> {
@@ -289,27 +392,37 @@ impl TryFrom<OwnedValue> for UserSelf {
     }
 }
 
-/// The type of Session. If `State::Invalid` then the response from
+impl IntoUserPath for UserPath {
+    fn into_path(&self) -> OwnedObjectPath {
+        self.path.clone()
+    }
+
+    fn into_path_ref(&self) -> &OwnedObjectPath {
+        &self.path
+    }
+}
+
+/// State of a User. If `UserState::Invalid` then the response from
 /// logind was not well defined.
-#[derive(Debug, PartialEq)]
-pub enum SessionType {
-    X11,
-    Wayland,
-    MIR,
-    TTY,
-    Unspecified,
+#[derive(Debug, PartialEq, Clone, Copy, Serialize, Deserialize)]
+pub enum UserState {
+    Online,
+    Offline,
+    Lingering,
+    Active,
+    Closing,
     Invalid,
 }
 
-impl From<&str> for SessionType {
+impl From<&str> for UserState {
     fn from(s: &str) -> Self {
-        match s {
-            "wayland" => SessionType::Wayland,
-            "x11" => SessionType::X11,
-            "mir" => SessionType::MIR,
-            "tty" => SessionType::TTY,
-            "unspecified" => SessionType::Unspecified,
-            _ => SessionType::Invalid,
+        match s.trim() {
+            "online" => Self::Online,
+            "offline" => Self::Offline,
+            "lingering" => Self::Lingering,
+            "active" => Self::Active,
+            "closing" => Self::Closing,
+            _ => Self::Invalid,
         }
     }
 }